2 Hour Housecleaning
I changed a lot here so let me go through some of it.
I fixed the errors in Project4 that I caused.
It was something to do with a missing .ToList on an IEnumera*
I added READMEs to everything and I removed the project name so
people don't try to copy my code to cheat.